@@ -7,9 +7,9 @@
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^5.3.3",
     "emailjs-com": "^2.6.4",
-    "react": "^17.0.1",
+    "react": "^18.1.0",
     "react-bootstrap": "^2.10.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.23.0",
     "react-scripts": "^5.0.1",
     "smooth-scroll": "^16.1.3"
@@ -34,5 +34,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "web-vitals": "^3.5.2"
   }
 }
